feat: redisEvent -> redisService
