+++ services/apply.js
@@ -9,7 +9,6 @@ const { ApplyEnum } = require('../utils/variables/enum');
 
 const User = require('../models/user');
 const Room = require('../models/room');
-const redisEvent = require('../events/redis');
 
 const createApply = async ({ user_id, study_id, message }) => {
   const createdRows = await applyDao.createApply({ user_id, study_id, message });
@@ -84,8 +83,7 @@ const applyProcess = async ({ study_id, apply_id }, { allow }) => {
     Room.updateOne({ study_id }, { $addToSet: { members: user_id } }).exec();
     User.updateOne({ user_id }, { $addToSet: { rooms: study_id } }).exec();
 
-    redisEvent.emit('trigger', RedisEventEnum.participate, user_id, { study_id });
-
+    redisTrigger(user_id, RedisEventEnum.participate, { study_id });
     broadcast.participate(study_id, nickname);
   } else {
     const rejectRows = await applyDao.setReject(apply_id);
